+++ test/smoke/testfixture_test.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/opentable/sous/dev_support/sous_qa_setup/desc"
 	sous "github.com/opentable/sous/lib"
+	"github.com/samsalisbury/semv"
 )
 
 type TestFixture struct {
@@ -24,17 +25,32 @@ type TestFixture struct {
 	Parent        *ParallelTestFixture
 	TestName      string
 	UserEmail     string
+	Projects      ProjectList
 	knownToFail   bool
 }
 
-func newTestFixture(t *testing.T, parent *ParallelTestFixture, nextAddr func() string, fcfg fixtureConfig) TestFixture {
+type fixtureConfig struct {
+	dbPrimary  bool
+	startState *sous.State
+	projects   ProjectList
+}
+
+func (fcfg fixtureConfig) Desc() string {
+	store := "GIT"
+	if fcfg.dbPrimary {
+		store = "DB"
+	}
+	return fmt.Sprintf("%s/%s", store, fcfg.projects.GroupName)
+}
+
+var sousBin = mustGetSousBin()
+
+func newTestFixture(t *testing.T, envDesc desc.EnvDesc, parent *ParallelTestFixture, nextAddr func() string, fcfg fixtureConfig) *TestFixture {
 	t.Helper()
 	t.Parallel()
 	if testing.Short() {
 		t.Skipf("-short flag present")
 	}
-	sousBin := getSousBin(t)
-	envDesc := getEnvDesc(t)
 	baseDir := getDataDir(t)
 
 	clusterSuffix := strings.Replace(t.Name(), "/", "_", -1)
@@ -66,28 +82,63 @@ func newTestFixture(t *testing.T, parent *ParallelTestFixture, nextAddr func() s
 		t.Fatalf("starting test cluster: %s", err)
 	}
 
-	client := makeClient(baseDir, sousBin)
 	primaryServer := "http://" + c.Instances[0].Addr
 	userEmail := "sous_client1@example.com"
-	if err := client.Configure(primaryServer, envDesc.RegistryName(), userEmail); err != nil {
-		t.Fatal(err)
-	}
 
-	return TestFixture{
+	tf := &TestFixture{
 		Cluster:       *c,
-		Client:        client,
 		BaseDir:       baseDir,
 		Singularity:   singularity,
 		ClusterSuffix: clusterSuffix,
 		Parent:        parent,
 		TestName:      t.Name(),
 		UserEmail:     userEmail,
+		Projects:      fcfg.projects,
 	}
+	client := makeClient(tf, baseDir, sousBin)
+	if err := client.Configure(primaryServer, envDesc.RegistryName(), userEmail); err != nil {
+		t.Fatal(err)
+	}
+	tf.Client = client
+	return tf
 }
 
-func (f *TestFixture) Stop(t *testing.T) {
+func (f *TestFixture) DIDAndDefaultReqID(t *testing.T, repo, offset, flavor, cluster string) (sous.DeploymentID, string) {
+	did := sous.DeploymentID{
+		ManifestID: sous.ManifestID{
+			Source: sous.SourceLocation{
+				Repo: repo,
+				Dir:  offset,
+			},
+			Flavor: flavor,
+		},
+		Cluster: cluster,
+	}
+	return did, f.Singularity.DefaultReqID(t, did)
+}
+
+// IsolatedClusterName returns a cluster name unique to this test fixture.
+func (f *TestFixture) IsolatedClusterName(baseName string) string {
+	return baseName + f.ClusterSuffix
+}
+
+// IsolatedVersionTag returns an all-lowercase unique version tag (unique per
+// test-run, subsequent runs will use the same tag). These version tags should
+// be compatible natively as both Sous and Docker tags.
+func (f *TestFixture) IsolatedVersionTag(t *testing.T, baseTag string) string {
 	t.Helper()
-	f.Cluster.Stop(t)
+	v, err := semv.Parse(baseTag)
+	if err != nil {
+		t.Fatalf("version tag %q not semver: %s", baseTag, err)
+	}
+	if v.Meta != "" {
+		t.Fatalf("version tag %q contains metatdata field", baseTag)
+	}
+	suffix := strings.Replace(f.ClusterSuffix, "_", "-", -1)
+	if v.Pre != "" {
+		return strings.ToLower(baseTag + suffix)
+	}
+	return strings.ToLower(baseTag + "-" + suffix)
 }
 
 func (f *TestFixture) ReportStatus(t *testing.T) {